:wrench: All components separated
