+++ src/components/Option.js
@@ -8,3 +8,5 @@ const Option = (props) => {
         </div>
     );
 };
+
+export default Option;
